Package the relevant classes into package named "duke"
