+++ src/main/java/duke/Deadline.java
@@ -1,7 +1,10 @@
-public class Event extends Task {
-    public Event(String description) {
+package duke;
+
+public class Deadline extends Task {
+    public Deadline(String description) {
         super(description);
     }
+
     public String digitMonth_to_AlphabeticalMonth(String digitMonth) {
         if (digitMonth.equals("01")) {
             return "January";
@@ -34,9 +37,10 @@ public class Event extends Task {
     }
 
     public String getDescription() {
-        String newReply = super.description.replace("event ", "");
+
+        String newReply = super.description.replace("deadline ", "");
         String taskAtHand = newReply.split("/")[0];
-        String deadline = newReply.split("/")[1].replace("at ", "at: ");
+        String deadline = newReply.split("/")[1].replace("by ", "by: ");
         String date = deadline.split(" ")[1].split("-")[0];
         String month = digitMonth_to_AlphabeticalMonth(deadline.split(" ")[1].split("-")[1]);
         String year = deadline.split(" ")[1].split("-")[2];
@@ -46,8 +50,7 @@ public class Event extends Task {
         } else {
             time += "am";
         }
-        System.out.println(date + month + year);
-        String finalDescription = taskAtHand + "(" + "at " + date + " " + month + " " + year + ", " + time + ")";
-        return "[E]" + "[" + super.getStatusIcon() + "] " + finalDescription;
+        String finalDescription = taskAtHand + "(" + "by " + date + " " + month + " " + year + ", " + time + ")";
+        return "[D]" + "[" + super.getStatusIcon() + "] " + finalDescription;
     }
 }
