+++ src/main/java/duke/Outputs.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class Outputs {
 
     public Outputs() {
