+++ src/main/java/duke/Todo.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.Task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
